vaxcert do not truncate

+++ app/Console/Commands/CovidVaccineLinelistImporter.php
@@ -84,8 +84,10 @@ class CovidVaccineLinelistImporter extends Command
         }
         */
 
-        CovidVaccinePatientMasterlist::truncate();
-
+        if (count(File::allFiles('/path/to/folder')) != 0 && File::exists(storage_path('app/vaxcert/masterlist/1.xlsx'))) {
+            CovidVaccinePatientMasterlist::truncate();
+        }
+        
         foreach($filenames as $f) {
             if(File::exists($f)) {
                 ProcessCovidVaccineMasterlistLinelist::dispatch($f);
